Progress being made in connecting Kibana and Elasticsearch securely
Although the reso-certification project had a .p12 file, this was not for securing internal communication between docker nodes.  We need to do much more work to make this happen.

More work needs to be done before this will work without having the error:
"unable to verify the first certificate"

We don't want to bypass the initial first-time certificate verification